Moved BADBAD hack to shuffle's cpp file
